+++ include/cowel/builtin_directive_set.hpp
@@ -152,6 +152,28 @@ Char_Get_Num_Behavior final : Int_Directive_Behavior {
     Result<Integer, Processing_Status> do_evaluate(const Invocation&, Context&) const final;
 };
 
+enum struct Text_Transformation : Default_Underlying {
+    lowercase,
+    uppercase,
+};
+
+struct [[nodiscard]]
+Str_Transform_Behavior final : Directive_Behavior {
+private:
+    const Text_Transformation m_transform;
+
+public:
+    [[nodiscard]]
+    constexpr explicit Str_Transform_Behavior(Text_Transformation transform)
+        : Directive_Behavior { Type::str }
+        , m_transform { transform }
+    {
+    }
+
+    [[nodiscard]]
+    Result<Value, Processing_Status> evaluate(const Invocation&, Context&) const final;
+};
+
 // clang-format off
 inline constexpr std::u8string_view lorem_ipsum = u8"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
 // clang-format on