+++ src/main/cpp/builtin_directive_set.cpp
@@ -122,6 +122,10 @@ constexpr Policy_Behavior cowel_source_as_text //
     { Known_Content_Policy::source_as_text };
 constexpr Unary_Numeric_Expression_Behavior cowel_sqrt //
     { Unary_Numeric_Expression_Kind::sqrt };
+constexpr Str_Transform_Behavior cowel_str_to_lower //
+    { Text_Transformation::lowercase };
+constexpr Str_Transform_Behavior cowel_str_to_upper //
+    { Text_Transformation::uppercase };
 constexpr N_Ary_Numeric_Expression_Behavior cowel_sub //
     { N_Ary_Numeric_Expression_Kind::sub };
 constexpr Policy_Behavior cowel_text_as_html //
@@ -431,6 +435,8 @@ constexpr Name_And_Behavior behaviors_by_name[] {
     COWEL_NAME_AND_BEHAVIOR_ENTRY(cowel_rem_to_zero),
     COWEL_NAME_AND_BEHAVIOR_ENTRY(cowel_source_as_text),
     COWEL_NAME_AND_BEHAVIOR_ENTRY(cowel_sqrt),
+    COWEL_NAME_AND_BEHAVIOR_ENTRY(cowel_str_to_lower),
+    COWEL_NAME_AND_BEHAVIOR_ENTRY(cowel_str_to_upper),
     COWEL_NAME_AND_BEHAVIOR_ENTRY(cowel_sub),
     COWEL_NAME_AND_BEHAVIOR_ENTRY(cowel_text_as_html),
     COWEL_NAME_AND_BEHAVIOR_ENTRY(cowel_text_only),